feat(backend): organizing models, and adding receiver service logic
